Moved messages and constants to relevant files
