+++ tests/test_connection.py
@@ -1,49 +1,39 @@
+import hashlib
 import uuid
-from datetime import datetime
 
 import pytest
-
-from smbprotocol.constants import Commands, Ciphers, CloseFlags, CtlCode, \
-    Dialects, \
-    HashAlgorithms, IOCTLFlags, NegotiateContextType, SecurityMode, Smb1Flags2
-from smbprotocol.messages import DirectTCPPacket, SMB1NegotiateRequest, \
-    SMB1PacketHeader, SMB2CloseRequest, SMB2CloseResponse, \
-    SMB2EncryptionCapabilities, \
-    SMB2ErrorResponse, SMB2FlushRequest, SMB2FlushResponse, SMB2IOCTLRequest, \
-    SMB2IOCTLResponse, SMB2Logoff, \
+from cryptography.hazmat.primitives.ciphers import aead
+from datetime import datetime
+from smbprotocol.connection import Ciphers, Commands, Dialects, \
+    HashAlgorithms, NegotiateContextType, SecurityMode, Smb1Flags2, \
+    SMB1NegotiateRequest, SMB1PacketHeader, SMB2EncryptionCapabilities, \
     SMB2NegotiateContextRequest, SMB2NegotiateRequest, SMB2NegotiateResponse, \
-    SMB2PacketHeader, SMB2PreauthIntegrityCapabilities, \
-    SMB2SessionSetupRequest, SMB2SessionSetupResponse, SMB2TransformHeader, \
-    SMB2TreeConnectRequest, SMB2TreeConnectResponse, SMB2TreeDisconnect, \
-    SMB2ValidateNegotiateInfoRequest, SMB2ValidateNegotiateInfoResponse, \
+    SMB2PacketHeader, SMB2PreauthIntegrityCapabilities, SMB2TransformHeader, \
     SMB3NegotiateRequest, SMB3PacketHeader
 
 
-class TestDirectTcpPacket(object):
+def test_valid_hash_algorithm():
+    expected = hashlib.sha512
+    actual = HashAlgorithms.get_algorithm(0x1)
+    assert actual == expected
 
-    def test_create_message(self):
-        message = DirectTCPPacket()
-        message['smb2_message'] = b"\xfe\x53\x4d\x42"
-        expected = b"\x00\x00\x00\x04" \
-                   b"\xfe\x53\x4d\x42"
 
-        actual = message.pack()
-        assert len(message) == 8
-        assert message['stream_protocol_length'].get_value() == 4
-        assert actual == expected
+def test_invalid_hash_algorithm():
+    with pytest.raises(KeyError) as exc:
+        HashAlgorithms.get_algorithm(0x2)
+        assert False  # shouldn't be reached
 
-    def test_parse_message(self):
-        actual = DirectTCPPacket()
-        data = b"\x00\x00\x00\x04" \
-               b"\xfe\x53\x4d\x42"
-        actual.unpack(data)
-        assert len(actual) == 8
-        assert actual['stream_protocol_length'].get_value() == 4
-        assert isinstance(actual['smb2_message'].get_value(), bytes)
 
-        actual_header = actual['smb2_message']
-        assert len(actual_header) == 4
-        assert actual_header.get_value() == b"\xfe\x53\x4d\x42"
+def test_valid_cipher():
+    expected = aead.AESCCM
+    actual = Ciphers.get_cipher(0x1)
+    assert actual == expected
+
+
+def test_invalid_cipher():
+    with pytest.raises(KeyError) as exc:
+        Ciphers.get_cipher(0x3)
+        assert False  # shouldn't be reached
 
 
 class TestSMB1PacketHeader(object):
@@ -260,45 +250,6 @@ class TestSMB3PacketHeader(object):
         assert actual['session_id'].get_value() == 10
         assert actual['signature'].get_value() == b"\x00" * 16
         assert actual['data'].get_value() == b"\x01\x02\x03\x04"
-
-
-class TestSMB2ErrorResponse(object):
-
-    def test_create_message_plain(self):
-        # This is a plain error response without the error context response
-        # data appended
-        message = SMB2ErrorResponse()
-        expected = b"\x09\x00" \
-                   b"\x00" \
-                   b"\x00" \
-                   b"\x00\x00\x00\x00"
-        actual = message.pack()
-        assert len(actual) == 8
-        assert actual == expected
-
-    def test_parse_message_plain(self):
-        actual = SMB2ErrorResponse()
-        data = b"\x09\x00" \
-               b"\x00" \
-               b"\x00" \
-               b"\x00\x00\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 8
-        assert actual['structure_size'].get_value() == 9
-        assert actual['error_context_count'].get_value() == 0
-        assert actual['reserved'].get_value() == 0
-        assert actual['byte_count'].get_value() == 0
-        assert actual['error_data'].get_value() == []
-
-
-# TODO: Once I get proper responses from the server
-class TestSMB2ErrorContextResponse(object):
-
-    def test_create_message(self):
-        pass
-
-    def test_parse_message(self):
-        pass
 
 
 class TestSMB2NegotiateRequest(object):
@@ -552,7 +503,7 @@ class TestSMB2NegotiateContextRequest(object):
         with pytest.raises(Exception) as exc:
             actual.unpack(data)
         assert str(exc.value) == "Enum value 3 does not exist in enum type " \
-                                 "<class 'smbprotocol.constants." \
+                                 "<class 'smbprotocol.connection." \
                                  "NegotiateContextType'>"
 
 
@@ -862,598 +813,6 @@ class TestSMB2NegotiateResponse(object):
             HashAlgorithms.SHA_512
         ]
         assert preauth_cap['salt'].get_value() == b"\x22" * 32
-
-
-class TestSMB2SessionSetupRequest(object):
-
-    def test_create_message(self):
-        message = SMB2SessionSetupRequest()
-        message['security_mode'] = SecurityMode.SMB2_NEGOTIATE_SIGNING_ENABLED
-        message['buffer'] = b"\x01\x02\x03\x04"
-        expected = b"\x19\x00" \
-                   b"\x00" \
-                   b"\x01" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x58\x00" \
-                   b"\x04\x00" \
-                   b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-                   b"\x01\x02\x03\x04"
-        actual = message.pack()
-        assert len(message) == 28
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2SessionSetupRequest()
-        data = b"\x19\x00" \
-               b"\x00" \
-               b"\x01" \
-               b"\x00\x00\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x58\x00" \
-               b"\x04\x00" \
-               b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-               b"\x01\x02\x03\x04"
-        actual.unpack(data)
-        assert len(actual) == 28
-        assert actual['structure_size'].get_value() == 25
-        assert actual['flags'].get_value() == 0
-        assert actual['security_mode'].get_value() == 1
-        assert actual['capabilities'].get_value() == 0
-        assert actual['security_buffer_offset'].get_value() == 88
-        assert actual['security_buffer_length'].get_value() == 4
-        assert actual['previous_session_id'].get_value() == 0
-        assert actual['buffer'].get_value() == b"\x01\x02\x03\x04"
-
-
-class TestSMB2SessionSetupResponse(object):
-
-    def test_create_message(self):
-        message = SMB2SessionSetupResponse()
-        message['session_flags'] = 1
-        message['buffer'] = b"\x04\x03\x02\x01"
-        expected = b"\x09\x00" \
-                   b"\x01\x00" \
-                   b"\x48\x00" \
-                   b"\x04\x00" \
-                   b"\x04\x03\x02\x01"
-        actual = message.pack()
-        assert len(message) == 12
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2SessionSetupResponse()
-        data = b"\x09\x00" \
-               b"\x01\x00" \
-               b"\x48\x00" \
-               b"\x04\x00" \
-               b"\x04\x03\x02\x01"
-        actual.unpack(data)
-        assert len(actual) == 12
-        assert actual['structure_size'].get_value() == 9
-        assert actual['session_flags'].get_value() == 1
-        assert actual['security_buffer_offset'].get_value() == 72
-        assert actual['security_buffer_length'].get_value() == 4
-        assert actual['buffer'].get_value() == b"\x04\x03\x02\x01"
-
-
-class TestSMB2Logoff(object):
-
-    def test_create_message(self):
-        message = SMB2Logoff()
-        expected = b"\x04\x00" \
-                   b"\x00\x00"
-        actual = message.pack()
-        assert len(message) == 4
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2Logoff()
-        data = b"\x04\x00" \
-               b"\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 4
-        assert actual['structure_size'].get_value() == 4
-        assert actual['reserved'].get_value() == 0
-
-
-class TestSMB2TreeConnectRequest(object):
-
-    def test_create_message(self):
-        message = SMB2TreeConnectRequest()
-        message['flags'] = 2
-        message['buffer'] = "\\\\127.0.0.1\\c$".encode("utf-16-le")
-        expected = b"\x09\x00" \
-                   b"\x02\x00" \
-                   b"\x48\x00" \
-                   b"\x1c\x00" \
-                   b"\x5c\x00\x5c\x00\x31\x00\x32\x00" \
-                   b"\x37\x00\x2e\x00\x30\x00\x2e\x00" \
-                   b"\x30\x00\x2e\x00\x31\x00\x5c\x00" \
-                   b"\x63\x00\x24\x00"
-        actual = message.pack()
-        assert len(message) == 36
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2TreeConnectRequest()
-        data = b"\x09\x00" \
-               b"\x02\x00" \
-               b"\x48\x00" \
-               b"\x1c\x00" \
-               b"\x5c\x00\x5c\x00\x31\x00\x32\x00" \
-               b"\x37\x00\x2e\x00\x30\x00\x2e\x00" \
-               b"\x30\x00\x2e\x00\x31\x00\x5c\x00" \
-               b"\x63\x00\x24\x00"
-        actual.unpack(data)
-        assert len(actual) == 36
-        assert actual['structure_size'].get_value() == 9
-        assert actual['flags'].get_value() == 2
-        assert actual['path_offset'].get_value() == 72
-        assert actual['path_length'].get_value() == 28
-        assert actual['buffer'].get_value() == "\\\\127.0.0.1\\c$"\
-            .encode("utf-16-le")
-
-
-class TestSMB2TreeConnectResponse(object):
-
-    def test_create_message(self):
-        message = SMB2TreeConnectResponse()
-        message['share_type'] = 1
-        message['share_flags'] = 2
-        message['capabilities'] = 8
-        message['maximal_access'] = 10
-        expected = b"\x10\x00" \
-                   b"\x01" \
-                   b"\x00" \
-                   b"\x02\x00\x00\x00" \
-                   b"\x08\x00\x00\x00" \
-                   b"\x0a\x00\x00\x00"
-        actual = message.pack()
-        assert len(message) == 16
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2TreeConnectResponse()
-        data = b"\x10\x00" \
-               b"\x01" \
-               b"\x00" \
-               b"\x02\x00\x00\x00" \
-               b"\x08\x00\x00\x00" \
-               b"\x0a\x00\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 16
-        assert actual['structure_size'].get_value() == 16
-        assert actual['share_type'].get_value() == 1
-        assert actual['reserved'].get_value() == 0
-        assert actual['share_flags'].get_value() == 2
-        assert actual['capabilities'].get_value() == 8
-        assert actual['maximal_access'].get_value() == 10
-
-
-class TestSMB2TreeDisconnect(object):
-
-    def test_create_message(self):
-        message = SMB2TreeDisconnect()
-        expected = b"\x04\x00" \
-                   b"\x00\x00"
-        actual = message.pack()
-        assert len(message) == 4
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2TreeDisconnect()
-        data = b"\x04\x00" \
-               b"\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 4
-        assert actual['structure_size'].get_value() == 4
-        assert actual['reserved'].get_value() == 0
-
-
-class TestSMB2CreateRequest(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2CreateContextRequest(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2CreateResponse(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2CloseRequest(object):
-
-    def test_create_message(self):
-        message = SMB2CloseRequest()
-        message['flags'].set_flag(CloseFlags.SMB2_CLOSE_FLAG_POSTQUERY_ATTRIB)
-        message['file_id'] = b"\xff" * 16
-        expected = b"\x18\x00" \
-                   b"\x01\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff"
-        actual = message.pack()
-        assert len(actual) == 24
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2CloseRequest()
-        data = b"\x18\x00" \
-               b"\x01\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff"
-        actual.unpack(data)
-        assert len(actual) == 24
-        assert actual['structure_size'].get_value() == 24
-        assert actual['flags'].get_value() == \
-            CloseFlags.SMB2_CLOSE_FLAG_POSTQUERY_ATTRIB
-        assert actual['reserved'].get_value() == 0
-        assert actual['file_id'].get_value().pack() == b"\xff" * 16
-
-
-class TestSMB2CloseResponse(object):
-
-    def test_create_message(self):
-        message = SMB2CloseResponse()
-        message['creation_time'] = datetime.utcfromtimestamp(0)
-        message['last_access_time'] = datetime.utcfromtimestamp(0)
-        message['last_write_time'] = datetime.utcfromtimestamp(0)
-        message['change_time'] = datetime.utcfromtimestamp(0)
-        expected = b"\x3c\x00" \
-                   b"\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-                   b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-                   b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-                   b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-                   b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00"
-        actual = message.pack()
-        assert len(actual) == 60
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2CloseResponse()
-        data = b"\x3c\x00" \
-               b"\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-               b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-               b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-               b"\x00\x80\x3E\xD5\xDE\xB1\x9D\x01" \
-               b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-               b"\x00\x00\x00\x00\x00\x00\x00\x00" \
-               b"\x00\x00\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 60
-        assert actual['structure_size'].get_value() == 60
-        assert actual['flags'].get_value() == 0
-        assert actual['reserved'].get_value() == 0
-        assert actual['creation_time'].get_value() == \
-            datetime.utcfromtimestamp(0)
-        assert actual['last_access_time'].get_value() == \
-            datetime.utcfromtimestamp(0)
-        assert actual['last_write_time'].get_value() == \
-            datetime.utcfromtimestamp(0)
-        assert actual['change_time'].get_value() == \
-            datetime.utcfromtimestamp(0)
-        assert actual['allocation_size'].get_value() == 0
-        assert actual['end_of_file'].get_value() == 0
-        assert actual['file_attributes'].get_value() == 0
-
-
-class TestSMB2FlushRequest(object):
-
-    def test_create_message(self):
-        message = SMB2FlushRequest()
-        message['file_id'] = b"\xff" * 16
-        expected = b"\x18\x00" \
-                   b"\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff"
-        actual = message.pack()
-        assert len(message) == 24
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2FlushRequest()
-        data = b"\x18\x00" \
-               b"\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff"
-        actual.unpack(data)
-        assert len(actual) == 24
-        assert actual['structure_size'].get_value() == 24
-        assert actual['reserved1'].get_value() == 0
-        assert actual['reserved2'].get_value() == 0
-        assert actual['file_id'].pack() == b"\xff" * 16
-
-
-class TestSMB2FlushResponse(object):
-
-    def test_create_message(self):
-        message = SMB2FlushResponse()
-        expected = b"\x04\x00" \
-                   b"\x00\x00"
-        actual = message.pack()
-        assert len(message) == 4
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2FlushResponse()
-        data = b"\x04\x00" \
-               b"\x00\x00"
-        actual.unpack(data)
-        assert len(actual) == 4
-        assert actual['structure_size'].get_value() == 4
-        assert actual['reserved'].get_value() == 0
-
-
-class TestSMB2ReadRequest(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2ReadResponse(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2WriteRequest(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2WriteResponse(object):
-
-    def test_create_message(self):
-        a = ""
-
-    def test_parse_message(self):
-        a = ""
-
-
-class TestSMB2IOCTLRequest(object):
-
-    def test_create_message(self):
-        message = SMB2IOCTLRequest()
-        message['ctl_code'] = CtlCode.FSCTL_VALIDATE_NEGOTIATE_INFO
-        message['file_id'] = b"\xff" * 16
-        message['max_input_response'] = 12
-        message['max_output_response'] = 12
-        message['flags'] = IOCTLFlags.SMB2_0_IOCTL_IS_FSCTL
-        message['buffer'] = b"\x12\x13\x14\x15"
-        expected = b"\x39\x00" \
-                   b"\x00\x00" \
-                   b"\x04\x02\x14\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\x78\x00\x00\x00" \
-                   b"\x04\x00\x00\x00" \
-                   b"\x0c\x00\x00\x00" \
-                   b"\x78\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x0c\x00\x00\x00" \
-                   b"\x01\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x12\x13\x14\x15"
-        actual = message.pack()
-        assert len(message) == 60
-        assert actual == expected
-
-    def test_create_message_no_buffer(self):
-        message = SMB2IOCTLRequest()
-        message['ctl_code'] = CtlCode.FSCTL_VALIDATE_NEGOTIATE_INFO
-        message['file_id'] = b"\xff" * 16
-        message['flags'] = IOCTLFlags.SMB2_0_IOCTL_IS_FSCTL
-        expected = b"\x39\x00" \
-                   b"\x00\x00" \
-                   b"\x04\x02\x14\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x01\x00\x00\x00" \
-                   b"\x00\x00\x00\x00"
-        actual = message.pack()
-        assert len(message) == 56
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2IOCTLRequest()
-        data = b"\x39\x00" \
-               b"\x00\x00" \
-               b"\x04\x02\x14\x00" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\x78\x00\x00\x00" \
-               b"\x04\x00\x00\x00" \
-               b"\x0c\x00\x00\x00" \
-               b"\x78\x00\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x0c\x00\x00\x00" \
-               b"\x01\x00\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x12\x13\x14\x15"
-        actual.unpack(data)
-        assert len(actual) == 60
-        assert actual['structure_size'].get_value() == 57
-        assert actual['reserved'].get_value() == 0
-        assert actual['ctl_code'].get_value() == \
-            CtlCode.FSCTL_VALIDATE_NEGOTIATE_INFO
-        assert actual['file_id'].pack() == b"\xff" * 16
-        assert actual['input_offset'].get_value() == 120
-        assert actual['input_count'].get_value() == 4
-        assert actual['max_input_response'].get_value() == 12
-        assert actual['output_offset'].get_value() == 120
-        assert actual['output_count'].get_value() == 0
-        assert actual['max_output_response'].get_value() == 12
-        assert actual['flags'].get_value() == IOCTLFlags.SMB2_0_IOCTL_IS_FSCTL
-        assert actual['reserved2'].get_value() == 0
-        assert actual['buffer'].get_value() == b"\x12\x13\x14\x15"
-
-
-class TestSMB2ValidateNegotiateInfoRequest(object):
-
-    def test_create_message(self):
-        message = SMB2ValidateNegotiateInfoRequest()
-        message['capabilities'] = 8
-        message['guid'] = b"\x11" * 16
-        message['security_mode'] = 1
-        message['dialect_count'] = 2
-        message['dialects'] = [Dialects.SMB_2_0_2, Dialects.SMB_2_1_0]
-        expected = b"\x08\x00\x00\x00" \
-                   b"\x11\x11\x11\x11\x11\x11\x11\x11" \
-                   b"\x11\x11\x11\x11\x11\x11\x11\x11" \
-                   b"\x01\x00" \
-                   b"\x02\x00" \
-                   b"\x02\x02\x10\x02"
-        actual = message.pack()
-        assert len(message) == 28
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2ValidateNegotiateInfoRequest()
-        data = b"\x08\x00\x00\x00" \
-               b"\x11\x11\x11\x11\x11\x11\x11\x11" \
-               b"\x11\x11\x11\x11\x11\x11\x11\x11" \
-               b"\x01\x00" \
-               b"\x02\x00" \
-               b"\x02\x02\x10\x02"
-        actual.unpack(data)
-        assert len(actual) == 28
-        assert actual['capabilities'].get_value() == 8
-        assert actual['guid'].get_value() == uuid.UUID(bytes=b"\x11" * 16)
-        assert actual['security_mode'].get_value() == 1
-        assert actual['dialect_count'].get_value() == 2
-        assert actual['dialects'][0] == 514
-        assert actual['dialects'][1] == 528
-        assert len(actual['dialects'].get_value()) == 2
-
-
-class TestSMB2IOCTLResponse(object):
-
-    def test_create_message(self):
-        message = SMB2IOCTLResponse()
-        message['ctl_code'] = CtlCode.FSCTL_VALIDATE_NEGOTIATE_INFO
-        message['file_id'] = b"\xff" * 16
-        message['input_offset'] = 0
-        message['input_count'] = 0
-        message['output_offset'] = 112
-        message['output_count'] = 4
-        message['flags'] = IOCTLFlags.SMB2_0_IOCTL_IS_FSCTL
-        message['buffer'] = b"\x20\x21\x22\x23"
-        expected = b"\x31\x00\x00\x00" \
-                   b"\x04\x02\x14\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x70\x00\x00\x00" \
-                   b"\x04\x00\x00\x00" \
-                   b"\x01\x00\x00\x00" \
-                   b"\x00\x00\x00\x00" \
-                   b"\x20\x21\x22\x23"
-        actual = message.pack()
-        assert len(message) == 52
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2IOCTLResponse()
-        data = b"\x31\x00\x00\x00" \
-               b"\x04\x02\x14\x00" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\x00\x00\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x70\x00\x00\x00" \
-               b"\x04\x00\x00\x00" \
-               b"\x01\x00\x00\x00" \
-               b"\x00\x00\x00\x00" \
-               b"\x20\x21\x22\x23"
-        actual.unpack(data)
-        assert len(actual) == 52
-        assert actual['structure_size'].get_value() == 49
-        assert actual['reserved'].get_value() == 0
-        assert actual['ctl_code'].get_value() == \
-            CtlCode.FSCTL_VALIDATE_NEGOTIATE_INFO
-        assert actual['file_id'].pack() == b"\xff" * 16
-        assert actual['input_offset'].get_value() == 0
-        assert actual['input_count'].get_value() == 0
-        assert actual['output_offset'].get_value() == 112
-        assert actual['output_count'].get_value() == 4
-        assert actual['flags'].get_value() == IOCTLFlags.SMB2_0_IOCTL_IS_FSCTL
-        assert actual['reserved2'].get_value() == 0
-        assert actual['buffer'].get_value() == b"\x20\x21\x22\x23"
-
-
-class TestSMB2ValidateNegotiateInfoResponse(object):
-
-    def test_create_message(self):
-        message = SMB2ValidateNegotiateInfoResponse()
-        message['capabilities'] = 8
-        message['guid'] = b"\xff" * 16
-        message['security_mode'] = 0
-        message['dialect'] = Dialects.SMB_3_0_2
-        expected = b"\x08\x00\x00\x00" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-                   b"\x00\x00" \
-                   b"\x02\x03"
-        actual = message.pack()
-        assert len(message) == 24
-        assert actual == expected
-
-    def test_parse_message(self):
-        actual = SMB2ValidateNegotiateInfoResponse()
-        data = b"\x08\x00\x00\x00" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\xff\xff\xff\xff\xff\xff\xff\xff" \
-               b"\x00\x00" \
-               b"\x02\x03"
-        actual.unpack(data)
-        assert len(actual) == 24
-        assert actual['capabilities'].get_value() == 8
-        assert actual['guid'].get_value() == uuid.UUID(bytes=b"\xff" * 16)
-        assert actual['security_mode'].get_value() == 0
-        assert actual['dialect'].get_value() == Dialects.SMB_3_0_2
 
 
 class TestSMB2TransformHeader(object):
